+++ src/CDataFile.h
@@ -143,7 +143,7 @@ t_Str	GetNextWord(t_Str& CommandLine);
 int		CompareNoCase(t_Str str1, t_Str str2);
 void	Trim(t_Str& szStr);
 template <typename... Args>
-int     WriteLn(std::fstream& stream, std::format_string<Args...> fmt, Args &&... args);
+size_t  WriteLn(std::fstream& stream, std::format_string<Args...> fmt, Args &&... args);
 
 
 /// Class Definitions ///////////////////////////////////////////////////////////
